Search on Audits, with some refactoring
