+++ backend/dissemination/searchlib/search_major_program.py
@@ -7,6 +7,17 @@ import logging
 logger = logging.getLogger(__name__)
 
 
+def audit_search_major_program(params):
+    q = Q()
+    major_program_fields = params.get("major_program")
+    if "True" in major_program_fields:
+        q |= Q(is_major_program=True)
+    elif "False" in major_program_fields:
+        q |= Q(is_major_program=False)
+    return q
+
+
+# TODO: Update Post SOC Launch -> Remove unused
 def search_major_program(general_results, params):
     """
     Searches on FederalAward columns 'is_major'. Comes in as True/False, to search on Y/N.
