+++ backend/dissemination/searchlib/search_passthrough_name.py
@@ -7,6 +7,18 @@ import logging
 logger = logging.getLogger(__name__)
 
 
+def audit_search_passthrough_name(params):
+    q = Q()
+    passthrough_names = params.get("passthrough_name")
+    for term in passthrough_names:
+        q_sub = Q()
+        for sub in term.split():
+            q_sub.add(Q(passthrough_names__icontains=sub), Q.AND)
+        q.add(q_sub, Q.OR)
+    return q
+
+
+# TODO: Update Post SOC Launch -> Remove unused
 def search_passthrough_name(general_results, params):
     """
     Searches on Passthrough field 'passthrough_name'.
