+++ backend/dissemination/urls.py
@@ -33,4 +33,5 @@ urlpatterns = [
     path("search/", views.Search.as_view(), name="Search"),
     path("search/advanced/", views.AdvancedSearch.as_view(), name="AdvancedSearch"),
     path("summary/<str:report_id>", views.AuditSummaryView.as_view(), name="Summary"),
+    path("search/beta/", views.AuditSearch.as_view(), name="BetaSearch"),
 ]
